Move role management into separate module
