+++ cartridge.lua
@@ -26,6 +26,7 @@ local rpc = require('cartridge.rpc')
 local vars = require('cartridge.vars').new('cartridge')
 local auth = require('cartridge.auth')
 local utils = require('cartridge.utils')
+local roles = require('cartridge.roles')
 local admin = require('cartridge.admin')
 local webui = require('cartridge.webui')
 local argparse = require('cartridge.argparse')
@@ -431,7 +432,7 @@ local function cfg(opts, box_opts)
     end
 
     for _, role in ipairs(opts.roles or {}) do
-        local ok, err = confapplier.register_role(role)
+        local ok, err = roles.register_role(role)
         if not ok then
             return nil, err
         end
