+++ cartridge/vshard-utils.lua
@@ -8,6 +8,7 @@ local errors = require('errors')
 local vars = require('cartridge.vars').new('cartridge.vshard-utils')
 local pool = require('cartridge.pool')
 local utils = require('cartridge.utils')
+local roles = require('cartridge.roles')
 local topology = require('cartridge.topology')
 local twophase = require('cartridge.twophase')
 local confapplier = require('cartridge.confapplier')
@@ -40,7 +41,7 @@ local function validate_group_weights(group_name, topology)
             'replicasets[%s].weight must be non-negative, got %s', replicaset_uuid, replicaset.weight
         )
 
-        local enabled_roles = confapplier.get_enabled_roles(replicaset.roles)
+        local enabled_roles = roles.get_enabled_roles(replicaset.roles)
         if enabled_roles['vshard-storage'] and (replicaset.vshard_group or 'default') == group_name then
             num_storages = num_storages + 1
             total_weight = total_weight + (replicaset.weight or 0)
@@ -278,7 +279,7 @@ end
 -- @local
 -- @treturn {[string]=table,...}
 local function get_known_groups()
-    local known_roles = confapplier.get_known_roles()
+    local known_roles = roles.get_known_roles()
     if utils.table_find(known_roles, 'vshard-router') == nil then
         return {}
     end
@@ -405,7 +406,7 @@ local function can_bootstrap_group(group_name, vsgroup)
 end
 
 local function can_bootstrap()
-    local known_roles = confapplier.get_known_roles()
+    local known_roles = roles.get_known_roles()
     if utils.table_find(known_roles, 'vshard-router') == nil then
         return false
     end
